+++ tfe/resource_tfe_organization_token_test.go
@@ -99,6 +99,67 @@ func TestAccTFEOrganizationToken_existsWithForce(t *testing.T) {
 	})
 }
 
+func TestAccTFEOrganizationToken_existsWithoutExpiry(t *testing.T) {
+	token := &tfe.OrganizationToken{}
+	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+	orgName := fmt.Sprintf("tst-terraform-%d", rInt)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckTFEOrganizationTokenDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccTFEOrganizationToken_basic(rInt),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckTFEOrganizationTokenExists(
+						"tfe_organization_token.foobar", token),
+					resource.TestCheckResourceAttr(
+						"tfe_organization_token.foobar", "organization", orgName),
+				),
+			},
+
+			{
+				Config:      testAccTFEOrganizationToken_existsWithoutExpiry(rInt),
+				ExpectError: regexp.MustCompile(`must be a valid date or time`),
+			},
+		},
+	})
+}
+
+func TestAccTFEOrganizationToken_existsWithExpiry(t *testing.T) {
+	token := &tfe.OrganizationToken{}
+	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+	orgName := fmt.Sprintf("tst-terraform-%d", rInt)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckTFEOrganizationTokenDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccTFEOrganizationToken_basic(rInt),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckTFEOrganizationTokenExists(
+						"tfe_organization_token.foobar", token),
+					resource.TestCheckResourceAttr(
+						"tfe_organization_token.foobar", "organization", orgName),
+				),
+			},
+
+			{
+				Config: testAccTFEOrganizationToken_existsWithExpiry(rInt),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckTFEOrganizationTokenExists(
+						"tfe_organization_token.expiry", token),
+					resource.TestCheckResourceAttr(
+						"tfe_organization_token.expiry", "organization", orgName),
+				),
+			},
+		},
+	})
+}
+
 func TestAccTFEOrganizationToken_import(t *testing.T) {
 	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
 
@@ -213,5 +274,38 @@ resource "tfe_organization_token" "foobar" {
 resource "tfe_organization_token" "regenerated" {
   organization     = tfe_organization.foobar.id
   force_regenerate = true
+}`, rInt)
+}
+
+func testAccTFEOrganizationToken_existsWithoutExpiry(rInt int) string {
+	return fmt.Sprintf(`
+resource "tfe_organization" "foobar" {
+  name  = "tst-terraform-%d"
+  email = "admin@company.com"
+}
+
+resource "tfe_organization_token" "foobar" {
+  organization = tfe_organization.foobar.id
+}
+
+resource "tfe_organization_token" "error" {
+  organization = tfe_organization.foobar.id
+}`, rInt)
+}
+
+func testAccTFEOrganizationToken_existsWithExpiry(rInt int) string {
+	return fmt.Sprintf(`
+resource "tfe_organization" "foobar" {
+  name  = "tst-terraform-%d"
+  email = "admin@company.com"
+}
+
+resource "tfe_organization_token" "foobar" {
+  organization = tfe_organization.foobar.id
+}
+
+resource "tfe_organization_token" "expiry" {
+  organization  = tfe_organization.foobar.id
+  expired_at 	= "04/11/2051"
 }`, rInt)
 }